+++ src-tauri/src/http_request.rs
@@ -24,13 +24,14 @@ use tokio::fs::{create_dir_all, File};
 use tokio::io::AsyncWriteExt;
 use tokio::sync::watch::Receiver;
 use tokio::sync::{oneshot, Mutex};
+use yaak_models::manager::QueryManagerExt;
 use yaak_models::models::{
     Cookie, CookieJar, Environment, HttpRequest, HttpResponse, HttpResponseHeader,
     HttpResponseState, ProxySetting, ProxySettingAuth,
 };
 use yaak_models::queries::{
-    get_base_environment, get_http_response, get_or_create_settings, get_workspace,
-    update_response_if_id, upsert_cookie_jar, UpdateSource,
+    get_base_environment, get_http_response, get_or_create_settings, update_response_if_id,
+    upsert_cookie_jar, UpdateSource,
 };
 use yaak_plugins::events::{
     CallHttpAuthenticationRequest, HttpHeader, RenderPurpose, WindowContext,
@@ -48,7 +49,7 @@ pub async fn send_http_request<R: Runtime>(
 ) -> Result<HttpResponse> {
     let app_handle = window.app_handle().clone();
     let plugin_manager = app_handle.state::<PluginManager>();
-    let workspace = get_workspace(&app_handle, &unrendered_request.workspace_id).await?;
+    let workspace = window.queries().connect()?.get_workspace(&unrendered_request.workspace_id)?;
     let base_environment =
         get_base_environment(&app_handle, &unrendered_request.workspace_id).await?;
     let settings = get_or_create_settings(&app_handle).await;
