+++ src-tauri/yaak-models/Cargo.toml
@@ -1,5 +1,6 @@
 [package]
 name = "yaak-models"
+links = "yaak-models"
 version = "0.1.0"
 edition = "2021"
 publish = false
@@ -19,3 +20,6 @@ sqlx = { version = "0.8.0", default-features = false, features = ["migrate", "sq
 tauri = { workspace = true }
 thiserror = "2.0.11"
 ts-rs = { workspace = true, features = ["chrono-impl", "serde-json-impl"] }
+
+[build-dependencies]
+tauri-plugin = { workspace = true, features = ["build"] }
