+++ src-tauri/yaak-models/build.rs
@@ -1,0 +1,5 @@
+const COMMANDS: &[&str] = &["upsert", "delete"];
+
+fn main() {
+    tauri_plugin::Builder::new(COMMANDS).build();
+}
