+++ src-tauri/yaak-models/permissions/autogenerated/commands/upsert.toml
@@ -1,0 +1,13 @@
+# Automatically generated - DO NOT EDIT!
+
+"$schema" = "../../schemas/schema.json"
+
+[[permission]]
+identifier = "allow-upsert"
+description = "Enables the upsert command without any pre-configured scope."
+commands.allow = ["upsert"]
+
+[[permission]]
+identifier = "deny-upsert"
+description = "Denies the upsert command without any pre-configured scope."
+commands.deny = ["upsert"]
