+++ src-tauri/yaak-models/permissions/autogenerated/reference.md
@@ -1,0 +1,120 @@
+## Default Permission
+
+Default permissions for the plugin
+
+- `allow-upsert`
+- `allow-delete`
+
+## Permission Table
+
+<table>
+<tr>
+<th>Identifier</th>
+<th>Description</th>
+</tr>
+
+
+<tr>
+<td>
+
+`yaak-models:allow-delete`
+
+</td>
+<td>
+
+Enables the delete command without any pre-configured scope.
+
+</td>
+</tr>
+
+<tr>
+<td>
+
+`yaak-models:deny-delete`
+
+</td>
+<td>
+
+Denies the delete command without any pre-configured scope.
+
+</td>
+</tr>
+
+<tr>
+<td>
+
+`yaak-models:allow-delete-model`
+
+</td>
+<td>
+
+Enables the delete_model command without any pre-configured scope.
+
+</td>
+</tr>
+
+<tr>
+<td>
+
+`yaak-models:deny-delete-model`
+
+</td>
+<td>
+
+Denies the delete_model command without any pre-configured scope.
+
+</td>
+</tr>
+
+<tr>
+<td>
+
+`yaak-models:allow-upsert`
+
+</td>
+<td>
+
+Enables the upsert command without any pre-configured scope.
+
+</td>
+</tr>
+
+<tr>
+<td>
+
+`yaak-models:deny-upsert`
+
+</td>
+<td>
+
+Denies the upsert command without any pre-configured scope.
+
+</td>
+</tr>
+
+<tr>
+<td>
+
+`yaak-models:allow-upsert-model`
+
+</td>
+<td>
+
+Enables the upsert_model command without any pre-configured scope.
+
+</td>
+</tr>
+
+<tr>
+<td>
+
+`yaak-models:deny-upsert-model`
+
+</td>
+<td>
+
+Denies the upsert_model command without any pre-configured scope.
+
+</td>
+</tr>
+</table>
