+++ src-tauri/yaak-models/permissions/default.toml
@@ -1,0 +1,6 @@
+[default]
+description = "Default permissions for the plugin"
+permissions = [
+    "allow-upsert",
+    "allow-delete",
+]
